refactor: factor out thermometer slider, mouse over cursor change
