+++ src/components/lab/thermometer/Scale.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import { Group, Text, Rect, RegularPolygon } from 'react-konva';
+import { Group, Text, Rect } from 'react-konva';
 import {
   THERMOMETER_HEIGHT,
   THERMOMETER_WIDTH,
@@ -14,9 +14,8 @@ import {
   THERMOMETER_COLOR,
   THERMOMETER_STROKE_COLOR,
   SCALE_MAX_NUMBER_TICKS,
-  SLIDER_FILL_COLOR,
 } from '../../../config/constants';
-import { setTemperature } from '../../../actions';
+import Slider from './Slider';
 
 const temperatureToHeight = ({
   currentTemperature,
@@ -34,33 +33,8 @@ const temperatureToHeight = ({
   return ((-minTemperature + value) * heightBetweenTicks) / tickStep;
 };
 
-const heightToTemperature = ({
-  height,
-  heightBetweenTicks,
-  tickStep,
-  minTemperature,
-  maxTemperature,
-  offsetY,
-  thermometerHeight,
-}) => {
-  const newTemperature =
-    ((thermometerHeight + offsetY - height) * tickStep) / heightBetweenTicks +
-    minTemperature;
-
-  // clamp value
-  let value = newTemperature;
-  if (value < minTemperature) {
-    value = minTemperature;
-  } else if (value > maxTemperature) {
-    value = maxTemperature;
-  }
-
-  return newTemperature;
-};
-
 const Scale = ({
   currentTemperature,
-  dispatchSetTemperature,
   scales: { from, to },
   thermometerHeight,
   offsetY,
@@ -78,19 +52,19 @@ const Scale = ({
 
   // compute scale array at every step
   const nbScale = (roundTo - roundFrom) / tickStep + 1; // +1 to include max
-  let scales = Array.from(
+  let scaleArray = Array.from(
     { length: nbScale },
     (key, value) => value * tickStep + roundFrom,
   );
 
   // select marks at most number of scale we can display
   const maxNbScale = Math.floor(thermometerHeight / SCALE_HEIGHT);
-  const prop = Math.ceil(scales.length / maxNbScale);
+  const prop = Math.ceil(scaleArray.length / maxNbScale);
   if (prop > 1) {
-    scales = scales.filter((_, i) => i % prop === 0);
+    scaleArray = scaleArray.filter((_, i) => i % prop === 0);
   }
 
-  const heightBetweenTicks = thermometerHeight / scales.length;
+  const heightBetweenTicks = thermometerHeight / scaleArray.length;
 
   // compute fill height given current temperature value
   const fillValue = temperatureToHeight({
@@ -102,7 +76,7 @@ const Scale = ({
   });
 
   // draw scale ticks
-  const scaleComponents = scales.map((text, idx) => {
+  const scaleComponents = scaleArray.map((text, idx) => {
     const thermometerYPosition =
       offsetY + thermometerHeight - idx * heightBetweenTicks;
 
@@ -125,16 +99,6 @@ const Scale = ({
     );
   });
 
-  const sliderPositionX =
-    THERMOMETER_POSITION_X +
-    THERMOMETER_WIDTH +
-    SCALE_WIDTH +
-    SCALE_PADDING_LEFT +
-    SCALE_FONT_SIZE * 2;
-
-  const minThermometerHeight = offsetY + thermometerHeight;
-  const maxThermomerterHeight =
-    minThermometerHeight - (scales.length - 1) * heightBetweenTicks;
   const currentTemperatureY = offsetY + thermometerHeight - fillValue;
   return (
     <>
@@ -151,47 +115,20 @@ const Scale = ({
       {scaleComponents}
 
       {/* triangle slider */}
-      <RegularPolygon
-        draggable
-        dragBoundFunc={(pos) => {
-          // clamp y position
-          let newPositionY = pos.y;
-          if (newPositionY > minThermometerHeight) {
-            newPositionY = minThermometerHeight;
-          } else if (newPositionY < maxThermomerterHeight) {
-            newPositionY = maxThermomerterHeight;
-          }
-          // compute temperature from slider y position
-          const newTemperature = heightToTemperature({
-            heightBetweenTicks,
-            height: newPositionY,
-            tickStep,
-            minTemperature: roundFrom,
-            maxTemperature: roundTo,
-            offsetY,
-            thermometerHeight,
-          });
-
-          dispatchSetTemperature(newTemperature);
-
-          return {
-            x: sliderPositionX,
-            y: newPositionY,
-          };
-        }}
-        x={sliderPositionX}
+      <Slider
+        heightBetweenTicks={heightBetweenTicks}
         y={currentTemperatureY}
-        sides={3}
-        radius={8}
-        rotation={30}
-        fill={SLIDER_FILL_COLOR}
+        offsetY={offsetY}
+        thermometerHeight={thermometerHeight}
+        minTemperature={roundFrom}
+        maxTemperature={roundTo}
+        tickStep={tickStep}
       />
     </>
   );
 };
 
 Scale.propTypes = {
-  dispatchSetTemperature: PropTypes.func.isRequired,
   currentTemperature: PropTypes.number.isRequired,
   scales: PropTypes.shape({
     from: PropTypes.number.isRequired,
@@ -210,8 +147,4 @@ const mapStateToProps = ({ lab }) => ({
   scales: lab.scales,
 });
 
-const mapDispatchToProps = {
-  dispatchSetTemperature: setTemperature,
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(Scale);
+export default connect(mapStateToProps)(Scale);
